Allow passing of custom sleep implementation
Either by passing a callable as the `sleeper` argument, or by inheriting
from `Tailer` and overwriting the `sleeper` method, one may provide
a custom implementation of the sleep call.

This may be useful in situations where additional abort logic is required,
like in a thread that should not block on a file that's not written to.